Language System transition seems to be done
Adjusted the way IPSignShop fetched the language from InventoryProfiles. 
Made a method to get that, InventoryProfiles now hides it's underlying
implementation.

It seems like the new language system is working. May be wrong though,
but tests suggested it.

Now pushing a stable state, so a reset is possible if I fail horribly at
changing the command system :P