Finished transition of the Command System for "InventoryProfiles"
Rewrote all commands to now use the "Tree Command System".
This won't directly affect this plugin, but I find it cleaner.

Made the commands reloadable by providing a public method in the main
class. This means you can change the language and then reload the
commands to reflect the updates in their keywords, names and patterns.

Deleted IPTabCompleter and CommandManager as well as CommandPreset
classes, as they are no longer needed.